pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptfuzz"
version = "0.1.0"
description = "Adversarial security testing framework for LLM applications"
readme = "README.md"
license = { text = "AGPL-3.0" }
authors = [{ name = "PromptFuzz Contributors" }]
requires-python = ">=3.10"
keywords = ["llm", "security", "fuzzing", "ai", "adversarial", "red-teaming"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "click>=8.0",
    "httpx>=0.24",
    "rich>=13.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "httpx>=0.24",
]

[project.scripts]
promptfuzz = "promptfuzz.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["promptfuzz*"]

[tool.setuptools.package-data]
promptfuzz = ["attacks/*.json"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"promptfuzz/reporter.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
